ENH : add make_scalp_surfaces function to public API
