+++ mne/commands/mne_make_scalp_surfaces.py
@@ -15,21 +15,11 @@ Examples
     $ mne make_scalp_surfaces --overwrite --subject sample
 
 """
-import copy
 import os
-import os.path as op
 import sys
 
 import mne
-from mne.utils import (run_subprocess, verbose, logger, ETSContext,
-                       get_subjects_dir)
-
-
-def _check_file(fname, overwrite):
-    """Prevent overwrites."""
-    if op.isfile(fname) and not overwrite:
-        raise IOError('File %s exists, use --overwrite to overwrite it'
-                      % fname)
+from mne.bem import make_scalp_surfaces
 
 
 def run():
@@ -59,79 +49,13 @@ def run():
     if subject is None or subjects_dir is None:
         parser.print_help()
         sys.exit(1)
-    _run(subjects_dir, subject, options.force, options.overwrite,
-         options.no_decimate, options.verbose)
-
-
-@verbose
-def _run(subjects_dir, subject, force, overwrite, no_decimate, verbose=None):
-    this_env = copy.deepcopy(os.environ)
-    subjects_dir = get_subjects_dir(subjects_dir, raise_error=True)
-    this_env['SUBJECTS_DIR'] = subjects_dir
-    this_env['SUBJECT'] = subject
-    if 'FREESURFER_HOME' not in this_env:
-        raise RuntimeError('The FreeSurfer environment needs to be set up '
-                           'for this script')
-    incomplete = 'warn' if force else 'raise'
-    subj_path = op.join(subjects_dir, subject)
-    if not op.exists(subj_path):
-        raise RuntimeError('%s does not exist. Please check your subject '
-                           'directory path.' % subj_path)
-
-    mri = 'T1.mgz' if op.exists(op.join(subj_path, 'mri', 'T1.mgz')) else 'T1'
-
-    logger.info('1. Creating a dense scalp tessellation with mkheadsurf...')
-
-    def check_seghead(surf_path=op.join(subj_path, 'surf')):
-        surf = None
-        for k in ['lh.seghead', 'lh.smseghead']:
-            this_surf = op.join(surf_path, k)
-            if op.exists(this_surf):
-                surf = this_surf
-                break
-        return surf
-
-    my_seghead = check_seghead()
-    if my_seghead is None:
-        run_subprocess(['mkheadsurf', '-subjid', subject, '-srcvol', mri],
-                       env=this_env)
-
-    surf = check_seghead()
-    if surf is None:
-        raise RuntimeError('mkheadsurf did not produce the standard output '
-                           'file.')
-
-    bem_dir = op.join(subjects_dir, subject, 'bem')
-    if not op.isdir(bem_dir):
-        os.mkdir(bem_dir)
-    dense_fname = op.join(bem_dir, '%s-head-dense.fif' % subject)
-    logger.info('2. Creating %s ...' % dense_fname)
-    _check_file(dense_fname, overwrite)
-    # Helpful message if we get a topology error
-    msg = '\n\nConsider using --force as an additional input parameter.'
-    surf = mne.bem._surfaces_to_bem(
-        [surf], [mne.io.constants.FIFF.FIFFV_BEM_SURF_ID_HEAD], [1],
-        incomplete=incomplete, extra=msg)[0]
-    mne.write_bem_surfaces(dense_fname, surf, overwrite=overwrite)
-    levels = 'medium', 'sparse'
-    tris = [] if no_decimate else [30000, 2500]
-    if os.getenv('_MNE_TESTING_SCALP', 'false') == 'true':
-        tris = [len(surf['tris'])]  # don't actually decimate
-    for ii, (n_tri, level) in enumerate(zip(tris, levels), 3):
-        logger.info('%i. Creating %s tessellation...' % (ii, level))
-        logger.info('%i.1 Decimating the dense tessellation...' % ii)
-        with ETSContext():
-            points, tris = mne.decimate_surface(points=surf['rr'],
-                                                triangles=surf['tris'],
-                                                n_triangles=n_tri)
-        dec_fname = dense_fname.replace('dense', level)
-        logger.info('%i.2 Creating %s' % (ii, dec_fname))
-        _check_file(dec_fname, overwrite)
-        dec_surf = mne.bem._surfaces_to_bem(
-            [dict(rr=points, tris=tris)],
-            [mne.io.constants.FIFF.FIFFV_BEM_SURF_ID_HEAD], [1], rescale=False,
-            incomplete=incomplete, extra=msg)
-        mne.write_bem_surfaces(dec_fname, dec_surf, overwrite=overwrite)
+    make_scalp_surfaces(
+        subject=subject,
+        subjects_dir=subjects_dir,
+        force=options.force,
+        overwrite=options.overwrite,
+        no_decimate=options.no_decimate,
+        verbose=options.verbose)
 
 
 mne.utils.run_command_if_main()
